Moved voxel graph tests to their own file
